refactor(config): relocate vendored_rules package and probe dataclasses to config/
Move the vendored validation-rules infrastructure from engines/ to
config/ so the generic @model_validator (landing in a follow-up PR) can
consume rules without crossing the engines-layer boundary.

Moved:
- src/llenergymeasure/engines/vendored_rules/ ->
  src/llenergymeasure/config/vendored_rules/ (loader + corpus JSON)
- DormantField, ConfigProbe dataclasses ->
  src/llenergymeasure/config/probe.py (from engines/protocol.py)
- tests/unit/engines/vendored_rules/ ->
  tests/unit/config/vendored_rules/

Path updates:
- config-rules-refresh.yml VENDORED_DIR
- scripts/vendor_rules.py, scripts/update_engine_rules.sh docstrings
  and output paths
- scripts/walkers/_base.py and configs/validation_rules/README.md
  references

Pure refactor: no behavioural change. All 138 rule-matching / corpus-
invariant / probe-config tests still pass; import-linter contracts
kept (foundation-independence now genuinely clean — no ignore_imports
required for the follow-up validator PR).
