+++ src/llenergymeasure/config/vendored_rules/__init__.py
@@ -2,15 +2,12 @@
 
 The corpus at ``configs/validation_rules/{engine}.yaml`` is the SSOT for
 what runtime config validation tells users about their configs. This package
-consumes that corpus and exposes a typed matcher API.
-
-This module does not yet wire into ``config/models.py`` — that wiring lands
-in phase 50.2c alongside the generic ``@model_validator``. Landing the
-loader contract first lets the vendor CI pipeline (50.2b) and the generic
-validator (50.2c) each depend on a stable surface.
+consumes that corpus and exposes a typed matcher API used by the generic
+``@model_validator`` on :class:`llenergymeasure.config.models.ExperimentConfig`
+and (eventually) by the sweep canonicaliser.
 """
 
-from llenergymeasure.engines.vendored_rules.loader import (
+from llenergymeasure.config.vendored_rules.loader import (
     VALID_ADDED_BY,
     VALID_EMISSION_CHANNEL,
     VALID_OUTCOME,
