+++ src/llenergymeasure/engines/tensorrt.py
@@ -27,7 +27,8 @@ from pathlib import Path
 from typing import Any
 
 from llenergymeasure.config.models import ExperimentConfig
-from llenergymeasure.engines.protocol import ConfigProbe, DormantField, InferenceOutput
+from llenergymeasure.config.probe import ConfigProbe, DormantField
+from llenergymeasure.engines.protocol import InferenceOutput
 from llenergymeasure.utils.exceptions import ConfigError, EngineError
 
 logger = logging.getLogger(__name__)
